+++ wwwroot/server-portal/manage/js/global-signalr.js
@@ -1,7 +1,6 @@
 ﻿document.addEventListener("DOMContentLoaded", () => {
     Gs.Variables.SignalR.streamChat = new signalR.HubConnectionBuilder()
         .withUrl("/StreamChat")
-        .withAutomaticReconnect()
         //.configureLogging(signalR.LogLevel.Information)
         .build();
 
@@ -9,18 +8,15 @@
     Gs.Variables.SignalR.streamChat.onclose(Gs.SignalR.Start());
 
 
-    Gs.Variables.SignalR.streamChat.on("ReceiveMessage", (user, message) => {
-        console.log("Receive Message", user, message);
-    });
-
-
     Gs.Variables.SignalR.streamChat.on("ReceiveStreamChat", (user, message) => {
-        Gs.Objects.OpenShareReceive();
-        setTimeout(function () { $("#StreamAdmin").html(user); }, 100);
+        if (message.targetWindow = "ShareReceivePanel") {
+            Gs.Objects.OpenShareReceive();
+            setTimeout(function () { $("#StreamAdmin").html(user); }, 100);
+        } else { Gs.Objects.OpenShareWindow(); }
 
-        message = JSON.parse(message);
+        message = JSON.parse(message).message;
         message.position = "left";
-        message.avatar = message.name == "Anonymous" ? "/server-portal/images/anonymous.jpg" : message.avatar;
+        message.avatar = message.name == "127.0.0.1" ? "/server-portal/images/anonymous.jpg" : message.avatar;
 
         let chatMessageList = Metro.storage.getItem("ShareChatMessageList", null);
         if (chatMessageList.filter(obj => { return obj.id == message.id }).length == 0) {
@@ -35,7 +31,7 @@
         } else {
             $("#ShareChatMessageCount").html((parseInt($("#ShareChatMessageCount").html()) + 1).toString());
         }
-        console.log("Receive Stream", user, JSON.parse(message));
+        //console.log("Receive Stream", user, JSON.parse(message));
     });
 
 
@@ -49,13 +45,17 @@
 
 
     Gs.Variables.SignalR.streamChat.on("ResponseUsers", (response) => {
-        let select = Metro.getPlugin("#userList", "select"); let options = []; select.data(""); let users = JSON.parse(response);
-        users.forEach(user => { if (Gs.Variables.username != user) { options.push({ val: user, title: user, selected: false }); } });
-        select.addOptions(options);  
-        //console.log("Users", response);
+        let select = Metro.getPlugin("#userList", "select");
+        if (select != undefined) {
+            let options = []; select.data(""); let users = JSON.parse(response);
+            users.forEach(user => { if (Gs.Variables.username != user) { options.push({ val: user, title: user, selected: false }); } });
+            select.addOptions(options);
+            //console.log("Users", response);
+        }
     });
 
     Gs.SignalR.Start();
+    setTimeout(function () { Gs.SignalR.GetUsers(); }, 1000);
 });
 
 
@@ -69,14 +69,8 @@ Gs.SignalR.Start = async function () {
 }
 
 
-Gs.SignalR.SendMessage = function (user, message) {
-    Gs.Variables.SignalR.streamChat.invoke("SendMessage", user, message).catch(function (err) {
-        return console.error(err);
-    });
-}
-
-
-Gs.SignalR.GetUsers = function () {
+Gs.SignalR.GetUsers =async function () {
+    if (Gs.Variables.SignalR.streamChat.state == signalR.HubConnectionState.Disconnected) { await Gs.Variables.SignalR.streamChat.start(); }
     if (Gs.Variables.SignalR.streamChat.state === signalR.HubConnectionState.Connected) {
         Gs.Variables.SignalR.streamChat.invoke("RequestUsers").catch(function (err) {
             return console.error(err);
@@ -87,13 +81,10 @@ Gs.SignalR.GetUsers = function () {
 //data:image/jpeg;base64,
 Gs.SignalR.StartVideoStream = async function (stopStreaming) {
     try {
-        if (Gs.Variables.SignalR.streamChat.state === signalR.HubConnectionState.Disconnected) {
-            await Gs.Variables.SignalR.streamChat.start();
-        }
+        if (Gs.Variables.SignalR.streamChat.state === signalR.HubConnectionState.Disconnected) { await Gs.Variables.SignalR.streamChat.start(); }
 
         let select = Metro.getPlugin("#userList", "select");
         let image = Gs.Media.GetVideoFrame("videoPreview");
-        //let message = JSON.stringify({ image: image.split(",")[1], message: "" });
         let message = JSON.stringify({ image: image, message: "" });
 
       
@@ -114,15 +105,30 @@ Gs.SignalR.StartVideoStream = async function (stopStreaming) {
 }
 
 
-Gs.SignalR.SendStreamChatMessage = function (message) {
+Gs.SignalR.SendStreamChatMessage = async function (message) {
+    if (Gs.Variables.SignalR.streamChat.state === signalR.HubConnectionState.Disconnected) { await Gs.Variables.SignalR.streamChat.start(); }
+
     let chatMessageList = Metro.storage.getItem("ShareChatMessageList", null);
-/*    Gs.Variables.chatmessage = message;*/
     chatMessageList.push(message);
     Metro.storage.setItem("ShareChatMessageList", chatMessageList);
 
-    let select = Metro.getPlugin("#userList", "select");
+    let targetUser;
+    if ($("#SharePanel").html() != undefined) {
+        let select = Metro.getPlugin("#userList", "select");
+        targetUser = select.val();
+        message.TargetUser = targetUser;
+        message.CallUser = Gs.Variables.username;
+        message.targetWindow = "ShareReceivePanel";
+    } else if ($("#ShareReceivePanel").html() != undefined) {
+        targetUser = $("#StreamAdmin").html();
+        message.TargetUser = targetUser;
+        message.CallUser = Gs.Variables.username;
+        message.targetWindow = "SharePanel";
+    }
+
+
     message = JSON.stringify({ message: message });
-    Gs.Variables.SignalR.streamChat.invoke("SendChatMessageToUser", select.val(), message).catch(function (err) {
+    Gs.Variables.SignalR.streamChat.invoke("SendChatMessageToUser", targetUser, message).catch(function (err) {
         return console.error(err);
     });
 }